Binary hardening: CFG, ASLR, CET, /sdl repo-wide + Dependabot action bumps (#53)
* feat(build): enable exploit mitigations repo-wide + action version bumps

Directory.Build.props applies to every Release binary: CFG (/guard:cf),
/sdl, ASLR, CET shadow stack (/CETCOMPAT), incremental linking off,
/OPT:REF,ICF. Removes the explicit RandomizedBaseAddress=false opt-outs
that were disabling ASLR on EIDCredentialProvider.dll and
EIDAuthenticationPackage.dll.

Local BinSkim: 49 findings -> 18. Remaining are BA2024 (Spectre:
requires the VS spectre-mitigated-libs component on build machines,
documented in the props file) and BA2027 (SourceLink, out of scope).

Also applies Dependabot bumps #49-#52: github-script v9, cache v6,
upload-artifact v7, download-artifact v8.


* feat(build): enable Spectre mitigations (/Qspectre) for Release builds

Requires the VS "Spectre-mitigated libs" component; the GitHub runner
image ships it and the dev machine now has it installed. Local BinSkim:
18 -> 10 findings. The one remaining BA2024 is an "unknown" verdict on
EIDAuthenticationPackage.dll caused by the toolset's own delayimp.lib
(pulled in by the deliberate advapi32 delay-load) shipping without
compiland debug info - not a real gap. The other 9 are BA2027
SourceLink, out of scope.


* feat(build): preflight check for Spectre-mitigated libs in build.ps1

Release builds now fail fast with install instructions (GUI + elevated
setup.exe command) when the VS Spectre libs component is missing,
instead of dying mid-build with MSB8040.


* fix(ci): pin C++ builds to windows-2025 for v143 Spectre libs

windows-latest now resolves to the vs2026 image; its VS 18 ships
Spectre-mitigated libs only for the v145 toolset, so v143 + /Qspectre
dies with MSB8040. windows-2025 carries VS 2022 with the v143 (14.44)
Spectre libs - same toolset as the dev machine. Pinned in windows-ci,
codeql (c-cpp leg), and scan-artifacts (build job only).

Also finishes the Dependabot bumps missed in the first pass:
download-artifact v8, cache v6, upload-artifact v7, github-script v9
in scan-artifacts.yml and aikido-badges.yml.


* fix(ci): windows-2022 is the only label that still carries VS 2022

The windows-2025 label also serves the vs2026 image now (verified in the
failed run: Image: windows-2025-vs2026), so the previous pin changed
nothing. windows-2022 is the remaining label with VS 2022 and the v143
Spectre-mitigated libs.


* fix: resolve all 7 CodeQL alerts + workflow hardening + SBOM in releases

Code (all 7 open CodeQL alerts on main, build-verified 10/10):
- EIDUserManagement: GetProfilesDirectoryW returns BOOL, not HRESULT -
  FAILED() never fired, so failure was silently ignored (real bug)
- StoredCredentialManagement: overflow guard + single-width DWORD buffer
  size shared by alloc/memset in the password encrypt path; USHORT-bound
  product for the decrypt path
- EIDSecuritySupportProvider: bound client-supplied UserLength/
  PasswordLength before any LSASS-heap size arithmetic; widen the
  SecureZeroMemory product
- Page02 wizard: replace sizeof-based pointer arithmetic with an
  explicit append offset
- MigrateUI: %zu for size_t, (int) cast for LRESULT list count

Workflows (Scorecard PinnedDependencies x23 + TokenPermissions x9):
- every action SHA-pinned with version comments
- default-deny top-level permissions: {} in all workflows
- SBOM: CI installs sbom-tool so build.ps1 generates the SPDX SBOM;
  shipped as EIDAuthentication-x64.sbom.zip in artifacts, release
  assets, and the provenance attestation


* refactor: resolve SonarCloud + Aikido quality findings on the PR

- auto for redundant type; C++17 init-statement scoping (SonarCloud)
- guard dwBlockLen == 0 after both CryptGetKeyParam calls before it is
  used as a divisor - the latent divide predated this branch but the
  new MAXDWORD/dwBlockLen guard made it visible (Aikido)
- pipeline Split-Path chain instead of nested calls in build.ps1
- Aikido nesting-depth note on the SSP bounds check declined: the guard
  sits at the same depth as its sibling validations in the SEH block


* refactor: auto for cast-typed declaration (SonarCloud S5827)

The previous fix moved the redundant-type smell one line down: with
auto on nUsed, the (DWORD) cast made the dwSize declaration type
redundant too.


* docs: design for working uninstaller EID certificate cleanup


* docs: implementation plan for uninstaller EID certificate cleanup


* feat: native EID certificate sweep for uninstall cleanup

RemoveAllEIDCertificates removes every cert with an EID: subject (the
wizard root CA) or EID: issuer (certs it issued) from LocalMachine
Root/CA/TrustedPeople/My and the same stores of every user profile,
mounting unloaded NTUSER.DAT hives. Deletes the CA machine key
container; machine-keyset check makes touching smart-card keys
impossible.


* feat: CleanupEIDCertificates rundll32 export for the uninstaller

Mirrors the CleanupLsaCredentials pattern; thin SEH-guarded wrapper
around RemoveAllEIDCertificates.


* fix: uninstaller certificate cleanup never ran and missed most stores

The old un.RemoveEIDCertificates passed its PowerShell script to
nsExec::ExecToLog as ~16 separate NSIS plugin arguments; ExecToLog pops
only the first, so PowerShell ran with an empty -Command and removed
nothing. Its filter also matched only EID: SUBJECTS, missing every
user certificate (issuer-only match), and skipped the LocalMachine
CA/TrustedPeople/My stores. Replaced with the native
CleanupEIDCertificates export. Cleanup checkboxes now default to
unchecked so an uninstall/reinstall cycle does not destroy enrollments.


* docs: add uninstaller certificate cleanup to VM test checklist


* fix(security): don't let a user-store certificate drive CA key deletion

Security review of the new cleanup sweep found a flaw I introduced in
b70f0a4. DeleteMachineKeyContainer took the container name, provider and
CRYPT_MACHINE_KEYSET flag straight out of a certificate's
CERT_KEY_PROV_INFO property and passed them to
CryptAcquireContext(CRYPT_DELETEKEYSET). That property is persisted
per-certificate and survives the registry round-trip, so any standard
user could silently add a cert with an "EID:" subject to their own
CurrentUser\CA or TrustedPeople store, name an arbitrary MACHINE key
container in it, and have the elevated uninstaller destroy that key.
The CRYPT_MACHINE_KEYSET check was not a safety rail - the attacker sets
that bit themselves. It only ever excluded the smart-card case.

Two checks that attacker-supplied data cannot forge now gate deletion:
- fMachineStore, threaded through CleanOpenStore, is TRUE only for the
  LocalMachine loop; user stores and mounted profile hives pass FALSE.
  A legitimate CA is never in a user store (MakeTrustedCertifcate sets
  CRYPT_MACHINE_KEYSET only for SAVEON_SYSTEMSTORE -> LocalMachine Root).
- ContainerKeyMatchesCertificate opens the container read-only and
  compares its public key against the certificate's SubjectPublicKeyInfo,
  so even a machine-store plant cannot misdirect the deletion.

Also from the review: treat NTE_BAD_KEYSET as benign (the CA cert in two
stores would otherwise log a spurious error), count and trace the
previously silent prov-info and mounted-hive failure paths, preserve
GetLastError across CloseHandle in EnablePrivilege, extract
SweepOneProfile to cut nesting, and resolve the 18 SonarCloud smells
(nullptr literals, auto, init-statements, range-for, std::array).
NSIS: real header text and a Create error check on the options page,
and an honest comment on what rundll32's exit code can actually report.


* refactor: C++17 init-statements for the last 4 SonarCloud S6004 smells

Clears the PR to zero open SonarCloud issues.


---------

+++ EIDMigrateUI/Page14_GroupSelect.cpp
@@ -60,7 +60,7 @@ INT_PTR CALLBACK WndProc_14_GroupSelect(HWND hwndDlg, UINT uMsg, WPARAM wParam, 
         // Update the count label
         WCHAR szCount[64];  // NOSONAR - LSASS-01: C-style buffer required by Win32 API
         swprintf_s(szCount, ARRAYSIZE(szCount), L"Selected: 0 of %d groups",
-            SendMessageW(hList, LB_GETCOUNT, 0, 0));
+            (int)SendMessageW(hList, LB_GETCOUNT, 0, 0));
         SetDlgItemText(hwndDlg, IDC_14_SELECTED_COUNT, szCount);
 
         return TRUE;
